requirements.txt
pyspark==3.5.1
kafka-python-ng==2.2.3
fastapi==0.128.8
uvicorn[standard]==0.39.0
pydantic==2.13.2
websockets==13.1
